New cukor

+++ server/models/inventory.go
@@ -8,7 +8,7 @@ import (
 type Inventory struct {
 	ID          int64           `json:"id" form:"id" gorm:"primarykey"`          // 库存ID
 	CommodityId int64           `json:"commodityId" form:"commodityId"`          // 水果ID
-	Quantity    int             `json:"quantity" form:"quantity"`                // 水果数量，箱
+	Quantity    float64         `json:"quantity" form:"quantity"`                // 水果数量，箱
 	EmployeeId  int64           `json:"employeeId" form:"employeeId"`            // 操作的员工ID
 	WarehouseId int64           `json:"warehouseId" form:"warehouseId"`          // 仓库ID
 	CreatedAt   time.Time       `json:"createdAt"`                               // 创建时间
